+++ src/main/scala/sangria/execution/ExecutionScheme.scala
@@ -94,4 +94,6 @@ case class ExecutionResult[Ctx, Res](
   ctx: Ctx,
   result: Res,
   errors: Vector[RegisteredError],
-  middlewareVals: List[(Any, Middleware[_])])
+  middlewareVals: List[(Any, Middleware[_])],
+  validationTiming: TimeMeasurement,
+  queryReducerTiming: TimeMeasurement)